Travis: Automatically test ALL debug code
This avoids silly problems like broken debug code. When debugging, I don't first
want to have to debug the debug code! :P

With all debug options enabled, `test-js` outputs 858081 lines. This is too much
for Travis, so we need to throw the debug output away on Travis. Note that
this still preserves information like *which* test failed, as well as the
precise error message.